+++ data-import-test-support/src/test/java/org/folio/rest/tools/utils/ModuleName.java
@@ -1,0 +1,12 @@
+package org.folio.rest.tools.utils;
+
+public class ModuleName {
+
+  public static String getModuleName() {
+    return "data_import_test_support";
+  }
+
+  public static String getModuleVersion() {
+    return "99999.0.0";
+  }
+}